backend/requirements.txt
flask==2.0.1
werkzeug==2.2.2
flask-cors==3.0.10
python-dotenv==0.19.0
google-generativeai==0.3.2
flask-jwt-extended==4.3.1
flask-sqlalchemy==2.5.1
bcrypt==4.0.1
python-multipart==0.0.5
requests==2.26.0
flask-socketio==5.3.6
pymongo==4.5.0
transformers==4.36.2
torch==2.6.0
python-engineio==4.7.1
python-socketio==5.9.0 